requirements.txt
alembic==1.16.4
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
Flask==2.2.5
Flask-Login==0.6.3
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.2.3
gspread==5.7.2
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
oauth2client==4.1.3
oauthlib==3.3.1
pandas==1.5.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.1.0
six==1.17.0
SQLAlchemy==1.4.50
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==2.2.3
